move useClasses
