+++ packages/utilities/etc/utilities.api.md
@@ -806,6 +806,20 @@ export { IStyleFunctionOrObject }
 // @public
 export function isVirtualElement(element: HTMLElement | IVirtualElement): element is IVirtualElement;
 
+// @public (undocumented)
+export interface IUseClassesOptions<TStylesProp, TStyleProps, TStyleSet> {
+    // (undocumented)
+    baseStyles: IStyleFunctionOrObject<TStyleProps, TStyleSet>;
+    // (undocumented)
+    customizationScopeName: string;
+    // (undocumented)
+    styleProps: Omit<TStyleProps, 'theme'>;
+    // (undocumented)
+    styles: TStylesProp;
+    // (undocumented)
+    useStaticStyles: boolean;
+}
+
 // @public
 export interface IVirtualElement extends HTMLElement {
     // (undocumented)
@@ -1191,6 +1205,11 @@ export const trProperties: string[];
 
 // @public
 export function unhoistMethods(source: any, methodNames: string[]): void;
+
+// @public (undocumented)
+export function useClasses<TStylesProp, TStyleProps extends {}, TStyleSet extends IStyleSet<TStyleSet>>(options: IUseClassesOptions<TStylesProp, TStyleProps, TStyleSet>): {
+    [key in keyof TStyleSet]: string;
+};
 
 // @public
 export function useCustomizationSettings(properties: string[], scopeName?: string, localSettings?: ICustomizations): ISettings;
